requirements.txt
attrs==19.1.0
beautifulsoup4==4.7.1
